+++ internal/dnsfilter/dnsfilter.go
@@ -148,17 +148,21 @@ const (
 	// FilteredBlockedService - the host is blocked by "blocked services" settings
 	FilteredBlockedService
 
-	// ReasonRewrite is returned when there was a rewrite by
-	// a legacy DNS Rewrite rule.
-	ReasonRewrite
+	// Rewritten is returned when there was a rewrite by a legacy DNS
+	// rewrite rule.
+	Rewritten
 
-	// RewriteAutoHosts is returned when there was a rewrite by
-	// autohosts rules (/etc/hosts and so on).
-	RewriteAutoHosts
+	// RewrittenAutoHosts is returned when there was a rewrite by autohosts
+	// rules (/etc/hosts and so on).
+	RewrittenAutoHosts
 
-	// DNSRewriteRule is returned when a $dnsrewrite filter rule was
-	// applied.
-	DNSRewriteRule
+	// RewrittenRule is returned when a $dnsrewrite filter rule was applied.
+	//
+	// TODO(a.garipov): Remove Rewritten and RewrittenAutoHosts by merging
+	// their functionality into RewrittenRule.
+	//
+	// See https://github.com/AdguardTeam/AdGuardHome/issues/2499.
+	RewrittenRule
 )
 
 // TODO(a.garipov): Resync with actual code names or replace completely
@@ -175,11 +179,9 @@ var reasonNames = []string{
 	FilteredSafeSearch:     "FilteredSafeSearch",
 	FilteredBlockedService: "FilteredBlockedService",
 
-	ReasonRewrite: "Rewrite",
-
-	RewriteAutoHosts: "RewriteEtcHosts",
-
-	DNSRewriteRule: "DNSRewriteRule",
+	Rewritten:          "Rewrite",
+	RewrittenAutoHosts: "RewriteEtcHosts",
+	RewrittenRule:      "RewriteRule",
 }
 
 func (r Reason) String() string {
@@ -331,15 +333,15 @@ type Result struct {
 	Rules []*ResultRule `json:",omitempty"`
 
 	// ReverseHosts is the reverse lookup rewrite result.  It is
-	// empty unless Reason is set to RewriteAutoHosts.
+	// empty unless Reason is set to RewrittenAutoHosts.
 	ReverseHosts []string `json:",omitempty"`
 
 	// IPList is the lookup rewrite result.  It is empty unless
-	// Reason is set to RewriteAutoHosts or ReasonRewrite.
+	// Reason is set to RewrittenAutoHosts or Rewritten.
 	IPList []net.IP `json:",omitempty"`
 
 	// CanonName is the CNAME value from the lookup rewrite result.
-	// It is empty unless Reason is set to ReasonRewrite.
+	// It is empty unless Reason is set to Rewritten or RewrittenRule.
 	CanonName string `json:",omitempty"`
 
 	// ServiceName is the name of the blocked service.  It is empty
@@ -379,7 +381,7 @@ func (d *DNSFilter) CheckHost(host string, qtype uint16, setts *RequestFiltering
 
 	// first - check rewrites, they have the highest priority
 	result = d.processRewrites(host, qtype)
-	if result.Reason == ReasonRewrite {
+	if result.Reason == Rewritten {
 		return result, nil
 	}
 
@@ -453,7 +455,7 @@ func (d *DNSFilter) CheckHost(host string, qtype uint16, setts *RequestFiltering
 func (d *DNSFilter) checkAutoHosts(host string, qtype uint16, result *Result) (matched bool) {
 	ips := d.Config.AutoHosts.Process(host, qtype)
 	if ips != nil {
-		result.Reason = RewriteAutoHosts
+		result.Reason = RewrittenAutoHosts
 		result.IPList = ips
 
 		return true
@@ -461,7 +463,7 @@ func (d *DNSFilter) checkAutoHosts(host string, qtype uint16, result *Result) (m
 
 	revHosts := d.Config.AutoHosts.ProcessReverse(host, qtype)
 	if len(revHosts) != 0 {
-		result.Reason = RewriteAutoHosts
+		result.Reason = RewrittenAutoHosts
 
 		// TODO(a.garipov): Optimize this with a buffer.
 		result.ReverseHosts = make([]string, len(revHosts))
@@ -488,7 +490,7 @@ func (d *DNSFilter) processRewrites(host string, qtype uint16) (res Result) {
 
 	rr := findRewrites(d.Rewrites, host)
 	if len(rr) != 0 {
-		res.Reason = ReasonRewrite
+		res.Reason = Rewritten
 	}
 
 	cnames := map[string]bool{}
@@ -696,7 +698,7 @@ func (d *DNSFilter) matchHost(host string, qtype uint16, setts RequestFilteringS
 	// awkward.
 	if dnsr := dnsres.DNSRewrites(); len(dnsr) > 0 {
 		res = d.processDNSRewrites(dnsr)
-		if res.Reason == DNSRewriteRule && res.CanonName == host {
+		if res.Reason == RewrittenRule && res.CanonName == host {
 			// A rewrite of a host to itself.  Go on and
 			// try matching other things.
 		} else {